# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateTagRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  model.ErrorResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      error:
        type: string
    type: object
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  model.LoginResponse:
    properties:
      email:
        type: string
      expires_at:
        type: string
      token:
        type: string
      user_id:
        type: integer
    type: object
...
  model.UpdateTagRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A REST API for personal budget management with transaction tracking,
    recurring payments, and reporting.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Budget API
  version: "1.0"
paths:
  /admin/run-scheduler:
    post:
      consumes:
      - application/json
      description: Manually trigger the scheduler to process recurring transactions
        due today
      produces:
      - application/json
      responses:
        "200":
          description: Scheduler execution result
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Run the scheduler
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, receive a session token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate the current session token
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
...
  /tags:
    get:
      consumes:
      - application/json
      description: Get all available tags for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of tags
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag for categorizing transactions
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/model.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new tag
      tags:
      - tags
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Tag deleted successfully
        "400":
          description: Invalid tag ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Tag not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a tag
      tags:
      - tags
    patch:
      consumes:
      - application/json
      description: Update an existing tag's name
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag update data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Tag not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a tag
      tags:
      - tags
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"
...